# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EcoTrack AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProfileDietType = typeof ProfileDietType[keyof typeof ProfileDietType];


export const ProfileDietType = {
  vegan: 'vegan',
  vegetarian: 'vegetarian',
  omnivore: 'omnivore',
} as const;

export interface Profile {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  dietType: ProfileDietType;
  streakDays: number;
  totalCarbonSaved: number;
  createdAt: string;
}

export type ProfileInputDietType = typeof ProfileInputDietType[keyof typeof ProfileInputDietType];


export const ProfileInputDietType = {
  vegan: 'vegan',
  vegetarian: 'vegetarian',
  omnivore: 'omnivore',
} as const;

export interface ProfileInput {
  name?: string;
  email?: string;
  dietType?: ProfileInputDietType;
  /** @nullable */
  avatarUrl?: string | null;
}

export type FootprintInputTransportMode = typeof FootprintInputTransportMode[keyof typeof FootprintInputTransportMode];


export const FootprintInputTransportMode = {
  car: 'car',
  bike: 'bike',
  bus: 'bus',
  train: 'train',
  walking: 'walking',
} as const;

export type FootprintInputDietType = typeof FootprintInputDietType[keyof typeof FootprintInputDietType];


export const FootprintInputDietType = {
  vegan: 'vegan',
  vegetarian: 'vegetarian',
  omnivore: 'omnivore',
} as const;

export interface FootprintInput {
  /** Daily km travelled */
  transportKm: number;
  transportMode: FootprintInputTransportMode;
  /** Monthly electricity usage in kWh */
  electricityKwh: number;
  dietType: FootprintInputDietType;
  /** Number of flights per year */
  flightsPerYear: number;
}

export type FootprintResultRating = typeof FootprintResultRating[keyof typeof FootprintResultRating];


export const FootprintResultRating = {
  excellent: 'excellent',
  good: 'good',
  average: 'average',
  poor: 'poor',
  critical: 'critical',
} as const;

export interface FootprintResult {
  totalKgCo2: number;
  transportKgCo2: number;
  electricityKgCo2: number;
  foodKgCo2: number;
  travelKgCo2: number;
  /** Score from 0-100, higher is greener */
  score: number;
  rating: FootprintResultRating;
}

export interface Entry {
  id: number;
  date: string;
  totalKgCo2: number;
  transportKgCo2: number;
  electricityKgCo2: number;
  foodKgCo2: number;
  travelKgCo2: number;
  transportMode: string;
  dietType: string;
  score: number;
}

export type EntryInputTransportMode = typeof EntryInputTransportMode[keyof typeof EntryInputTransportMode];


export const EntryInputTransportMode = {
  car: 'car',
  bike: 'bike',
  bus: 'bus',
  train: 'train',
  walking: 'walking',
} as const;

export type EntryInputDietType = typeof EntryInputDietType[keyof typeof EntryInputDietType];


export const EntryInputDietType = {
  vegan: 'vegan',
  vegetarian: 'vegetarian',
  omnivore: 'omnivore',
} as const;

export interface EntryInput {
  transportKm: number;
  transportMode: EntryInputTransportMode;
  electricityKwh: number;
  dietType: EntryInputDietType;
  flightsPerYear: number;
  /** @nullable */
  notes?: string | null;
}

export type RecommendationCategory = typeof RecommendationCategory[keyof typeof RecommendationCategory];


export const RecommendationCategory = {
  transport: 'transport',
  energy: 'energy',
  food: 'food',
  travel: 'travel',
  general: 'general',
} as const;

export type RecommendationDifficulty = typeof RecommendationDifficulty[keyof typeof RecommendationDifficulty];


export const RecommendationDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface Recommendation {
  id: number;
  category: RecommendationCategory;
  title: string;
  description: string;
  potentialSavingKg: number;
  difficulty: RecommendationDifficulty;
  priority: number;
}

export interface Badge {
  id: number;
  name: string;
  description: string;
  icon: string;
  category: string;
  earned: boolean;
  /** @nullable */
  earnedAt: string | null;
}

export type TipImpact = typeof TipImpact[keyof typeof TipImpact];


export const TipImpact = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Tip {
  id: number;
  category: string;
  title: string;
  body: string;
  impact: TipImpact;
}

export interface DashboardStats {
  currentMonthKgCo2: number;
  previousMonthKgCo2: number;
  percentChange: number;
  weeklyAvgKgCo2: number;
  carbonScore: number;
  streakDays: number;
  badgesEarned: number;
  totalEntries: number;
}

export interface TrendPoint {
  label: string;
  totalKgCo2: number;
  transportKgCo2: number;
  electricityKgCo2: number;
  foodKgCo2: number;
  travelKgCo2: number;
}

export interface BreakdownItem {
  category: string;
  kgCo2: number;
  percentage: number;
  color: string;
}

export type ListEntriesParams = {
/**
 * Filter by time period
 */
period?: ListEntriesPeriod;
};

export type ListEntriesPeriod = typeof ListEntriesPeriod[keyof typeof ListEntriesPeriod];


export const ListEntriesPeriod = {
  week: 'week',
  month: 'month',
  year: 'year',
} as const;

export type ListTipsParams = {
category?: ListTipsCategory;
};

export type ListTipsCategory = typeof ListTipsCategory[keyof typeof ListTipsCategory];


export const ListTipsCategory = {
  transport: 'transport',
  energy: 'energy',
  food: 'food',
  travel: 'travel',
  general: 'general',
} as const;

export type GetCarbonTrendParams = {
/**
 * Trend period
 */
period?: GetCarbonTrendPeriod;
};

export type GetCarbonTrendPeriod = typeof GetCarbonTrendPeriod[keyof typeof GetCarbonTrendPeriod];


export const GetCarbonTrendPeriod = {
  week: 'week',
  month: 'month',
} as const;