# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EcoTrack AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /entries:
    get:
      operationId: listEntries
      tags: [entries]
      summary: List all footprint log entries
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, year]
          description: Filter by time period
      responses:
        "200":
          description: List of entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entry"
    post:
      operationId: createEntry
      tags: [entries]
      summary: Log a new footprint entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntryInput"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
  /entries/{id}:
    delete:
      operationId: deleteEntry
      tags: [entries]
      summary: Delete a footprint entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Entry:
      type: object
      required: [id, date, totalKgCo2, transportKgCo2, electricityKgCo2, foodKgCo2, travelKgCo2, transportMode, dietType, score]
      properties:
        id:
          type: integer
        date:
          type: string
        totalKgCo2:
          type: number
        transportKgCo2:
          type: number
        electricityKgCo2:
          type: number
        foodKgCo2:
          type: number
        travelKgCo2:
          type: number
        transportMode:
          type: string
        dietType:
          type: string
        score:
          type: integer
    EntryInput:
      type: object
      required: [transportKm, transportMode, electricityKwh, dietType, flightsPerYear]
      properties:
        transportKm:
          type: number
        transportMode:
          type: string
          enum: [car, bike, bus, train, walking]
        electricityKwh:
          type: number
        dietType:
          type: string
          enum: [vegan, vegetarian, omnivore]
        flightsPerYear:
          type: integer
        notes:
          type: ["string", "null"]
...