# lib/api-zod/src/generated/types/entryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EcoTrack AI API
 * OpenAPI spec version: 0.1.0
 */
import type { EntryInputDietType } from './entryInputDietType';
import type { EntryInputTransportMode } from './entryInputTransportMode';

export interface EntryInput {
  transportKm: number;
  transportMode: EntryInputTransportMode;
  electricityKwh: number;
  dietType: EntryInputDietType;
  flightsPerYear: number;
  /** @nullable */
  notes?: string | null;
}